Create LambdaCheckStockRole.yaml
This CloudFormation template defines another IAM role named LambdaCheckStockRole. Let's analyze its components:

Type: AWS::IAM::Role specifies the AWS CloudFormation resource type.

Properties: This section contains the configuration details for the IAM role.

RoleName: Specifies the name of the IAM role as Lambda-Check-Stock-Role.

AssumeRolePolicyDocument: Defines who can assume (or switch to) this role. In this case, only the AWS service lambda.amazonaws.com is allowed to assume this role.

Version: The version of the policy document.

Statement: Defines the permissions policy. In this case, it allows Lambda service to assume the role.

Sid: Optional identifier for the statement.

Effect: Specifies whether the policy allows or denies the action.

Principal: Specifies the entity that is allowed or denied access. In this case, it's the Lambda service.

Action: Specifies the action that the principal can take. Here, it's sts:AssumeRole, which allows the Lambda service to assume the role.

ManagedPolicyArns: Specifies a list of managed policies to attach to the role. Managed policies are standalone policies that you can attach to multiple IAM roles, users, or groups in your AWS account.

arn:aws:iam::aws:policy/AmazonSNSFullAccess: This policy provides full access to Amazon SNS, allowing the Lambda function to perform any action on SNS resources.

arn:aws:iam::aws:policy/service-role/AWSLambdaDynamoDBExecutionRole: This policy grants the necessary permissions for Lambda functions to read records from DynamoDB Streams, which is typically used for stream-based processing of DynamoDB data.

This IAM role is intended to be assumed by Lambda functions responsible for checking stock levels in DynamoDB. It has permissions to interact with Amazon SNS for sending notifications and with DynamoDB for reading records.