Create LambdaLoadInventoryRole.yaml
This CloudFormation template defines an IAM role named LambdaLoadInventoryRole. Let's break down its components:

Type: AWS::IAM::Role specifies the AWS CloudFormation resource type.

Properties: This section contains the configuration details for the IAM role.

RoleName: Specifies the name of the IAM role as Lambda-Load-Inventory-Role.

AssumeRolePolicyDocument: Defines who can assume (or switch to) this role. In this case, only the AWS service lambda.amazonaws.com is allowed to assume this role. This is necessary because Lambda functions assume this role when they run.

Version: The version of the policy document.

Statement: Defines the permissions policy. In this case, it allows Lambda service to assume the role.

Effect: Specifies whether the policy allows or denies the action.

Principal: Specifies the entity that is allowed or denied access. In this case, it's the Lambda service.

Action: Specifies the action that the principal can take. Here, it's sts:AssumeRole, which allows the Lambda service to assume the role.

ManagedPolicyArns: Specifies a list of managed policies to attach to the role. Managed policies are standalone policies that you can attach to multiple IAM roles, users, or groups in your AWS account.

arn:aws:iam::aws:policy/AmazonS3ReadOnlyAccess: This policy provides read-only access to Amazon S3 buckets and objects.

arn:aws:iam::aws:policy/AmazonDynamoDBFullAccess: This policy provides full access to Amazon DynamoDB resources.

Policies: Specifies inline policies to attach to the role. Inline policies are policies that you create and manage directly within the role's definition.

PolicyName: Specifies the name of the inline policy as CWLogsPolicy.

PolicyDocument: Specifies the permissions granted by the inline policy.

Version: The version of the policy document.

Statement: Defines the permissions granted by the policy. Here, it allows the Lambda function to create log groups, create log streams, and put log events to CloudWatch Logs.

Action: Specifies the action that is allowed. Here, it includes actions related to CloudWatch Logs.

Resource: Specifies the AWS resource to which the action applies. In this case, it applies to all resources within CloudWatch Logs (arn:aws:logs:*:*:*).

Effect: Specifies whether the action is allowed (Allow).

This IAM role is intended to be assumed by Lambda functions responsible for loading inventory data into DynamoDB. It has permissions to read data from S3, write data to DynamoDB, and log events to CloudWatch Logs.